refactor: POST일 때, Body를 읽어 User를 생성하는 책임을 갖는새 객체 생성
